REF: Refactor linking into submodule
