+++ trackpy/linking/__init__.py
@@ -1,0 +1,4 @@
+from .linking import TreeFinder, link, link_iter, link_df_iter
+from .find_link import find_link, find_link_iter
+from .utils import verify_integrity, SubnetOversizeException, TrackUnstored, \
+                   Point, UnknownLinkingError
